porciones

+++ app/src/main/java/com/example/myapplication/Ingrediente.java
@@ -25,7 +25,7 @@ public class Ingrediente implements Serializable {
         return cantidad;
     }
 
-    public void setCantidad(int cantidad) {
+    public void setCantidad(double cantidad) {
         this.cantidad = cantidad;
     }
